+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.20
 
+* Remove type of button from `DefaultButton` and add companion to `UIKitButton` to be able to use `uk-button` class
+
 ## 0.0.19
 
 * Create `Card`s implementation